Changed the .gitignore to allow .sln
I realized that I needed to allow for .sln files for other projects (duh), I changed it so that the .sln file just doesn't allow for a specific solution in my dev branch.

I am using this dev branch as a 'direct' repo for the engine. This should only be used for stuff like the build system. Other engine changes can come from the projects that use them.