Fixed Component/Collider Registration Bug!
WHOOT!!!

There was a bug that causes the engine to crash if the static container that holds onto the factory for components and colliders, aren’t initialized before the components/colliders try to register themselves to that container. It was literally luck of the draw.

To fix this, instead of having a static member variable, I made a static function that returns a reference to a static container INSIDE that function. This will force the initialization if it hadn't happened already!