I now load AppProps from data! Updated Boilerplate code, Fixed Release mode.
I had a weird issue with different compiler versions between my projects. It seems to be down to a different version in one of the external libraries (I suspect Lua). I turned off an optimization setting in the Engine and Utility project and that seemed to rectify the code.

I now load the App Properties from data, similar to how I load the Log Categories. Very non-optimized, straightforward implementation.