Changing the Folder Structure, and Removing the Tools
I have been trying to get my build system up and running correctly and as a result, the Tools folder is going to live outside of the Engine.

I have made a script to generate necessary folders and template files for a new project using my engine (basically a primitive engine launcher). To get things to work nicely, I needed to remove the additional 'MCPEngine' folder. However, when my launcher downloads the engine, the engine is placed in a MCPEngine folder. Basically, I had a double situation. This is cleaner.