Created a Dependencies folder, updated Engine vcxproj
I created a Dependencies folder that houses the Utility and Lua projects. This now matches the solution folders that I created in the filters in vs. I updated the Additional includes directly (later should be done in the build script), and I updated the build script to account for the new locations.