Added Assert, Updated gitignore for ignoring Perforce files
Added _CHECK and _CHECK_MSG as my Asserts. (I didn't want to use _ASSERT because I think that could break things.)

Also added the Assert.h in MCPEngine.

Logger now creates a default "Assertion Failed!" Category. I removed it from the Enigne.config so that it can't be messed with. It needs to always be on.