Component Registry "Fix"
I have made the easy fix to the issue where Engine components not directly used in the project code and just loaded from data aren't being registered: I just made a header file that includes every engine component and added that to the Application.cpp file which will always be included in new projects.

Also fixed some whitespace in the Component.cpp file and moved the Logo into an Images folder in Assets.