Updated Message System API
It was very silly how the API was set up for Components to send messages to each other, so I made a simple interface in the component. This is much better.

The Message system now uses StringIds! Now we aren't hashing every. single. compare. Well, we are when accessing the Message Listeners  container, but we are only hashing a pointer, and we are no longer hashing to send a message.

I also got rid of the idea of the Messenger. Seemed like a weird extra step that wasn't necessary.

Also did a second pass at the stringId API, and made accessing the static map thread safe to future proof the class.