Collision Woes
Ok, I had an issue where Colliders were not being removed from their cells, causing all sorts of issues. That is fixed, but in a 'code-smelly' way. It is a hack, and it works. This needs more fleshing out.

Scene now has deferred deletion for the updateables. I had an updateable remove another during their update, which of course would cause issues. This also needs a fresh look.