Editor Config!
Fixed a bug where failing to initialize a system would result in a null reference exception when closing the application (trying to access the deleted Resource Manager).

Added "EditorDebug" and "Editor" configurations into the Engine. It is pretty hacky at the moment, but the Editor configuration allows you to edit a scene asset. This was very much geared toward completing the assignment, and not a series implementation of an Editor (that I really want to make).

- We can now save an XML file to it's source.